// web-rally/src/client/models/ActivityType.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * Enum for activity types with string values for database compatibility
 */
export enum ActivityType {
    TIME_BASED_ACTIVITY = 'TimeBasedActivity',
    SCORE_BASED_ACTIVITY = 'ScoreBasedActivity',
    BOOLEAN_ACTIVITY = 'BooleanActivity',
    TEAM_VS_ACTIVITY = 'TeamVsActivity',
    GENERAL_ACTIVITY = 'GeneralActivity',
}